apps/lambdas/trade/requirements.txt
boto3
alpaca-py
